+++ vitest.setup.ts
@@ -2,7 +2,6 @@ import { expect } from "vitest";
 import type { NewPlugin } from "@vitest/pretty-format";
 import { z } from "zod/v4";
 import type { $ZodType } from "zod/v4/core";
-import { InputValidationError, OutputValidationError } from "./src";
 import { isFunctionSchema } from "./src/function-schema";
 
 /** Takes cause and certain props of custom errors into account */
@@ -10,16 +9,10 @@ const errorSerializer: NewPlugin = {
   test: (subject) => subject instanceof Error,
   serialize: (error: Error, config, indentation, depth, refs, printer) => {
     const { name, message, cause } = error;
-    const { originalError } =
-      error instanceof InputValidationError ||
-      error instanceof OutputValidationError
-        ? error
-        : {};
     const { issues } = error instanceof z.ZodError ? error : {};
     const obj = Object.assign(
       issues ? { issues } : { message },
       cause && { cause },
-      originalError && { originalError },
     );
     return `${name}(${printer(obj, config, indentation, depth, refs)})`;
   },